Moved order logic to separate component
