+++ resume.js
@@ -1,51 +1,10 @@
-import Awards from './components/awards.js'
-import Certificates from './components/certificates.js'
-import Education from './components/education.js'
 import Header from './components/header.js'
-import Interests from './components/interests.js'
-import Languages from './components/languages.js'
 import Meta from './components/meta.js'
-import Projects from './components/projects.js'
-import Publications from './components/publications.js'
-import References from './components/references.js'
-import Skills from './components/skills.js'
-import Volunteer from './components/volunteer.js'
-import Work from './components/work.js'
 import colors from './utils/colors.js'
 import html from './utils/html.js'
+import Body from "./components/body.js";
 
 export default function Resume(resume, css) {
-  let resumeOrder = resume.meta.order
-  if (resumeOrder === undefined) {
-    resumeOrder = [
-      'work',
-      'volunteer',
-      'education',
-      'projects',
-      'awards',
-      'certificates',
-      'publications',
-      'skills',
-      'languages',
-      'interests',
-      'references',
-    ]
-  }
-
-  const orderComponentMap = {
-    work: Work(resume.work),
-    volunteer: Volunteer(resume.volunteer),
-    education: Education(resume.education),
-    projects: Projects(resume.projects),
-    awards: Awards(resume.awards),
-    certificates: Certificates(resume.certificates),
-    publications: Publications(resume.publications),
-    skills: Skills(resume.skills),
-    languages: Languages(resume.languages),
-    interests: Interests(resume.interests),
-    references: References(resume.references),
-  }
-
   return html`<!DOCTYPE html>
     <html lang="en" style="${colors(resume.meta)}">
       <head>
@@ -57,7 +16,7 @@ export default function Resume(resume, css) {
       </head>
       <body>
         ${Header(resume.basics)}
-        ${resumeOrder.map(section => html`${orderComponentMap[section]}`)}
+        ${Body(resume)}
       </body>
     </html>`
 }
